move buffer into file
