+++ lib/src/parsers/flac.dart
@@ -6,6 +6,7 @@ import 'package:audio_metadata_reader/src/metadata/base.dart';
 import 'package:audio_metadata_reader/src/metadata/vorbis_metadata.dart';
 import 'package:audio_metadata_reader/src/parsers/id3v2.dart';
 import 'package:audio_metadata_reader/src/parsers/tag_parser.dart';
+import 'package:audio_metadata_reader/src/utils/buffer.dart';
 
 import '../utils/bit_manipulator.dart';
